Refactoring to remove FakeBatchablePluginSinkBase
