+++ core/src/main/java/com/spotify/ffwd/util/BatchMetricConverter.java
