+++ modules/opencensus/pom.xml
@@ -22,7 +22,7 @@
 
     <dependency>
       <groupId>com.spotify.ffwd</groupId>
-      <artifactId>ffwd-api</artifactId>
+      <artifactId>ffwd-core</artifactId>
     </dependency>
 
     <dependency>
@@ -91,7 +91,11 @@
           </exclusion>
         </exclusions>
       </dependency>
-
+      <dependency>
+          <groupId>com.google.code.gson</groupId>
+          <artifactId>gson</artifactId>
+          <version>2.8.1</version>
+      </dependency>
 
   </dependencies>
 </project>
